fix(ewe-note): prefer healthy sync status
